requirements.txt
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.3.1
cssselect==1.3.0
curl_cffi==0.11.2
dnspython==2.7.0
Flask==3.1.2
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
lxml==5.4.0
MarkupSafe==3.0.3
packaging==25.0
parsel==1.10.0
psycopg2-binary==2.9.11
pycparser==2.22
pymongo==4.13.0
PyMySQL==1.1.2
python-dotenv==1.2.1
requests==2.32.3
SQLAlchemy==2.0.44
typing_extensions==4.15.0
urllib3==2.4.0
w3lib==2.3.1
Werkzeug==3.1.4
